ai-checker/requirements.txt
Flask==3.0.0
openai==1.12.0
mysql-connector-python==8.3.0
PyPDF2==3.0.1
python-docx==1.1.0